add blocklist for blocking the lohout access_tokens and impelement unit test for market

+++ resources/user.py
@@ -1,10 +1,18 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from passlib.hash import pbkdf2_sha256
-from flask_jwt_extended import create_access_token
+from flask_jwt_extended import (
+    create_access_token,
+    get_jwt,
+    jwt_required,
+    create_refresh_token,
+    get_jwt_identity
+)
 from db import db
+
 from models.user import UserModel
 from schemas import UserSchema
+from blocklist import BLOCKLIST
 
 
 
@@ -16,6 +24,7 @@ class UserRegister(MethodView):
     
     @blp.arguments(UserSchema)
     def post(self, user_info):
+        
         if UserModel.query.filter(UserModel.username==user_info["username"]).first():
             abort(409, message="A user with that username already exists.")
         
@@ -40,8 +49,13 @@ class User(MethodView):
         user = UserModel.query.get_or_404(user_id)
         return user
     
-
+    @jwt_required(refresh=True)
     def delete(self, user_id):
+        
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilage required.")
+            
         user = UserModel.query.get_or_404(user_id)
         db.session.delete(user)
         db.session.commit()
@@ -55,14 +69,47 @@ class login(MethodView):
     
     @blp.arguments(UserSchema)
     def post(self, user_info):
+    
         user = UserModel.query.filter(
             UserModel.username == user_info["username"]
         ).first()
         
         if user and pbkdf2_sha256.verify(user_info["password"], user.password):
-            access_token = create_access_token(identity=user.id)
+            access_token = create_access_token(identity=user.id, fresh=True)
+            refresh_token = create_refresh_token(user.id)
             
-            return {"access_tocken": access_token}, 200
+            return {"access_token": access_token, "refresh_token": refresh_token}, 200
         
         
         abort(401, message="Invalid credentials.")
+        
+        
+        
+    @blp.route("/logout")
+    class UserLogout(MethodView):
+        
+        @jwt_required()
+        def post(self):
+            
+            jti = get_jwt()['jti']
+            BLOCKLIST.add(jti)
+            
+            print(jti, "this is jti in logout processing")
+            print(BLOCKLIST, "this is blocklist set after user logout")
+            return {"message": "successfully logged out."}, 200
+    
+    
+    @blp.route("/refresh")
+    class TokenRefresh(MethodView):
+        
+        @jwt_required(refresh=True)
+        def post(self):
+            
+            current_user = get_jwt_identity()
+            new_token = create_access_token(identity=current_user, fresh=False)
+            jti = get_jwt()['jti']
+            BLOCKLIST.add(jti)
+            
+            print(jti, "this is jti in tocken refreshing processing")
+            print(BLOCKLIST, "this is blocklist set after token refreshing")
+            return {"access_token": new_token}, 200
